lua : move lua/widget.c to lua/widget/widget.c
